+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -6,14 +6,14 @@ from typing import Dict, Optional
 import yaml
 
 from demisto_sdk.commands.common.constants import (
-    BANG_COMMAND_ARGS_MAPPING_DICT, BANG_COMMAND_NAMES, DBOT_SCORES_DICT,
-    DEFAULT_CONTENT_ITEM_FROM_VERSION, DEPRECATED_REGEXES,
-    ENDPOINT_COMMAND_NAME, ENDPOINT_FLEXIBLE_REQUIRED_ARGS,
-    FEED_REQUIRED_PARAMS, FETCH_REQUIRED_PARAMS, FIRST_FETCH,
-    FIRST_FETCH_PARAM, INTEGRATION_CATEGORIES, IOC_OUTPUTS_DICT, MAX_FETCH,
-    MAX_FETCH_PARAM, PACKS_DIR, PACKS_PACK_META_FILE_NAME, PYTHON_SUBTYPES,
-    REPUTATION_COMMAND_NAMES, TYPE_PWSH, XSOAR_CONTEXT_STANDARD_URL,
-    XSOAR_SUPPORT)
+    ALERT_FETCH_REQUIRED_PARAMS, BANG_COMMAND_ARGS_MAPPING_DICT,
+    BANG_COMMAND_NAMES, DBOT_SCORES_DICT, DEFAULT_CONTENT_ITEM_FROM_VERSION,
+    DEPRECATED_REGEXES, ENDPOINT_COMMAND_NAME, ENDPOINT_FLEXIBLE_REQUIRED_ARGS,
+    FEED_REQUIRED_PARAMS, FIRST_FETCH, FIRST_FETCH_PARAM,
+    INCIDENT_FETCH_REQUIRED_PARAMS, INTEGRATION_CATEGORIES, IOC_OUTPUTS_DICT,
+    MAX_FETCH, MAX_FETCH_PARAM, PACKS_DIR, PACKS_PACK_META_FILE_NAME,
+    PYTHON_SUBTYPES, REPUTATION_COMMAND_NAMES, TYPE_PWSH,
+    XSOAR_CONTEXT_STANDARD_URL, XSOAR_SUPPORT, MarketplaceVersions)
 from demisto_sdk.commands.common.default_additional_info_loader import \
     load_default_additional_info_dict
 from demisto_sdk.commands.common.errors import (FOUND_FILES_AND_ERRORS,
@@ -28,8 +28,8 @@ from demisto_sdk.commands.common.hook_validations.docker import \
 from demisto_sdk.commands.common.hook_validations.image import ImageValidator
 from demisto_sdk.commands.common.tools import (
     _get_file_id, compare_context_path_in_yml_and_readme, get_core_pack_list,
-    get_file_version_suffix_if_exists, get_files_in_dir, get_pack_name,
-    is_iron_bank_pack, print_error, server_version_compare)
+    get_file_version_suffix_if_exists, get_files_in_dir, get_item_marketplaces,
+    get_pack_name, is_iron_bank_pack, print_error, server_version_compare)
 
 default_additional_info = load_default_additional_info_dict()
 
@@ -900,11 +900,18 @@ class IntegrationValidator(ContentEntityValidator):
         """
         fetch_params_exist = True
         if self.current_file.get('script', {}).get('isfetch') is True:
+
+            # get the iten marketplaces to decide which are the required params
+            # if no marketplaces or xsoar in marketplaces - the required params will be INCIDENT_FETCH_REQUIRED_PARAMS (with Incident type etc. )
+            # otherwise it will be the ALERT_FETCH_REQUIRED_PARAMS (with Alert type etc. )
+            marketplaces = get_item_marketplaces(item_path=self.file_path, item_data=self.current_file)
+            is_xsoar_marketplace = not marketplaces or MarketplaceVersions.XSOAR.value in marketplaces
+            fetch_required_params = INCIDENT_FETCH_REQUIRED_PARAMS if is_xsoar_marketplace else ALERT_FETCH_REQUIRED_PARAMS
             params = [dict.copy(_key) for _key in self.current_file.get('configuration', [])]
             for param in params:
                 if 'defaultvalue' in param:
                     param.pop('defaultvalue')
-            for param in FETCH_REQUIRED_PARAMS:
+            for param in fetch_required_params:
                 if param not in params:
                     error_message, error_code = Errors.parameter_missing_from_yml(param.get('name'),
                                                                                   yaml.dump(param))