+++ CHANGELOG.md
@@ -1,4 +1,27 @@
 # Changelog
+* Fixed an issue in the **create-id-set** command where similar items from different marketplaces were reported as duplicated.
+* Fixed typo in demisto-sdk init
+* Fixed an issue where the **lint** command did not handle all container exit codes.
+
+# 1.5.9
+* Added option to specify `External Playbook Configuration` to change inputs of Playbooks triggered as part of **test-content**
+* Improved performance of the **lint** command.
+* Improved performance of the **validate** command when checking README images.
+* ***create-id-set*** command - the default value of the **marketplace** argument was changed from ‘xsoar’ to all packs existing in the content repository. When using the command, make sure to pass the relevant marketplace to use.
+
+# 1.5.8
+* Fixed an issue where the command **doc-review** along with the argument `--release-notes` failed on yml/json files with invalid schema.
+* Fixed an issue where the **lint** command failed on packs using python 3.10
+
+# 1.5.7
+* Fixed an issue where reading remote yaml files failed.
+* Fixed an issue in **validate** failed with no error message for lists (when no fromVersion field was found).
+* Fixed an issue when running **validate** or **format** in a gitlab repository, and failing to determine its project id.
+* Added an enhancement to **split**, handling an empty output argument.
+* Added the ability to add classifiers and mappers to conf.json.
+* Added the Alias field to the incident field schema.
+
+# 1.5.6
 * Added 'deprecated' release notes template.
 * Fixed an issue where **run-test-playbook** command failed to get the task entries when the test playbook finished with errors.
 * Fixed an issue in **validate** command when running with `no-conf-json` argument to ignore the `conf.json` file.